Feature/156 session

+++ app/controllers/friends_controller.rb
@@ -5,9 +5,12 @@ class FriendsController < ApplicationController
   end
 
   def create
-    Friend.create(user_id: current_user.id, friend_user_id: params[:id])
-    Friend.create(user_id: params[:id],         friend_user_id: current_user.id)
-    request = Request.find_by(from_user_id: params[:id], to_user_id: current_user.id)
-    request.destroy
+    if request = Request.find_by(from_user_id: params[:id], to_user_id: current_user.id)
+      Friend.create(user_id: current_user.id, friend_user_id: params[:id])
+      Friend.create(user_id: params[:id],     friend_user_id: current_user.id)
+      request.destroy
+    else 
+      render json: { errors: [exception] }, status: 404
+    end
   end
 end
